Refactor: CategoryList분리
