+++ src/components/CategoryList/CategoryList.stories.tsx
